@@ -35,23 +35,23 @@
   },
   "dependencies": {
     "history": "^3.0.0",
-    "prop-types": "^15.5.8",
+    "prop-types": "^15.5.10",
     "react": "^15.5.4",
     "react-css-transition": "^0.7.4",
     "react-dom": "^15.5.4",
     "react-fastclick": "^3.0.1",
     "react-ga": "^2.2.0",
-    "react-helmet": "^5.0.2",
-    "react-lite": "^0.15.34",
+    "react-helmet": "^5.1.3",
+    "react-lite": "^0.15.37",
     "react-router": "3.0.0",
     "react-router-scroll": "^0.4.2",
-    "serve": "^5.1.2"
+    "serve": "^5.2.0"
   },
   "devDependencies": {
     "autoprefixer": "^6.7.7",
     "babel-cli": "^6.24 .0",
     "babel-core": "^6.24.0",
-    "babel-eslint": "^7.2.1",
+    "babel-eslint": "^7.2.3",
     "babel-loader": "^6.4.1",
     "babel-preset-es2015": "^6.24.0",
     "babel-preset-react": "^6.24.1",
@@ -59,26 +59,26 @@
     "compression-webpack-plugin": "^0.4.0",
     "copy-webpack-plugin": "^4.0.1",
     "cross-env": "^4.0.0",
-    "css-loader": "^0.28.0",
+    "css-loader": "^0.28.4",
     "eslint": "^3.19.0",
     "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-import": "^2.3.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.10.3",
     "expect": "^1.20.2",
-    "express": "^4.14.0",
+    "express": "^4.15.3",
     "extract-text-webpack-plugin": "^2.0.0-beta.5",
     "file-loader": "^0.11.1",
-    "image-webpack-loader": "^3.3.0",
+    "image-webpack-loader": "^3.3.1",
     "json-loader": "^0.5.4",
-    "karma": "^1.6.0",
-    "karma-chrome-launcher": "^2.0.0",
+    "karma": "^1.7.0",
+    "karma-chrome-launcher": "^2.1.1",
     "karma-mocha": "^1.3.0",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^2.0.2",
     "lodash": "^4.16.6",
     "mkdirp": "^0.5.1",
-    "mocha": "^3.2.0",
+    "mocha": "^3.4.2",
     "moment": "^2.18.1",
     "postcss-calc": "^5.3.1",
     "postcss-custom-media": "^5.0.1",
@@ -89,15 +89,15 @@
     "postcss-nested": "^1.0.1",
     "postcss-short": "^3.0.3",
     "react-addons-test-utils": "^15.5.1",
-    "redbox-react": "^1.3.6",
+    "redbox-react": "^1.3.7",
     "rimraf": "^2.5.4",
     "sitemap": "^1.12.0",
     "style-loader": "^0.16.1",
     "stylelint": "^7.10.1",
     "stylelint-config-standard": "^16.0.0",
     "url-loader": "^0.5.7",
-    "webpack": "^2.3.3",
-    "webpack-dev-middleware": "^1.8.4",
+    "webpack": "^2.6.1",
+    "webpack-dev-middleware": "^1.10.2",
     "webpack-hot-middleware": "^2.18.0",
     "webpack-isomorphic-tools": "^3.0.2"
   }
