@@ -1,7 +1,8 @@
 export default {
+  id: 2,
   name: 'Projects',
   description: 'Side projects',
-  url: '/projects',
+  url: '/projects/',
   parent: '/',
   type: 'drawer',
   items: [
