@@ -3,7 +3,7 @@ import image from '../../../client/images/work/ap.png';
 export default {
   name: 'Associated Press',
   description: 'Associated Press front-end code',
-  url: '/work/associated-press',
+  url: '/work/associated-press/',
   parent: '/work/',
   image: {
     url: image,
@@ -15,4 +15,5 @@ export default {
     'gulp.js', 'Jade', 'PostCSS',
   ],
   website: 'http://ap.org',
+  type: 'drawer',
 };
