@@ -3,7 +3,7 @@ import image from '../../../client/images/work/bauer.png';
 export default {
   name: 'Bauer Media',
   description: 'Bauer Media website',
-  url: '/work/bauer-media',
+  url: '/work/bauer-media/',
   parent: '/work/',
   image: {
     url: image,
@@ -16,4 +16,5 @@ export default {
   ],
   website: 'https://www.bauermedia.co.uk/',
   datePublished: '2013-10-12',
+  type: 'drawer',
 };
