@@ -1,24 +1,15 @@
+import work from '../work/';
+import projects from '../projects/';
+import about from '../about/';
+
 export default {
   name: '',
   alternateHeadline: 'Workbench',
   description: 'Web developer with over 12 years of professional experience',
   url: '/',
   items: [
-    {
-      id: '1',
-      name: 'Work',
-      url: '/work/',
-      type: 'drawer',
-    }, {
-      id: '2',
-      name: 'Projects',
-      url: '/projects/',
-      type: 'drawer',
-    }, {
-      id: '3',
-      name: 'About',
-      url: '/about/',
-      type: 'drawer',
-    },
+    work,
+    projects,
+    about,
   ],
 };
